third_party/WebKit/LayoutTests/virtual/presentation/receiver/presentation-receiver-sandboxing.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(t => {
  var result = window.open('about:blank', '_blank');
  assert_equals(result, null);
}, "Test that window.open() returns 'null' synchronously without blocking on user input.");

test(t => {
  var result = alert("Yay!");
  assert_equals(result, undefined);
}, "Test that in presentation receiver pages, alert() returns synchronously without blocking on user input.");

test(t => {
  var result = print();
  assert_equals(result, undefined);
}, "Test that in presentation receiver pages, print() returns synchronously without blocking on user input.");

test(t => {
  var result = confirm("Question?");
  assert_equals(result, false);
}, "Test that in presentation receiver pages, confirm() returns 'false' synchronously without blocking on user input.");

test(t => {
  var result = prompt("Question?");
  assert_equals(result, null);
}, "Test that in presentation receiver pages, prompt() returns 'null' synchronously without blocking on user input.");
</script>
</body>
</html>
